requirements.txt
Flask==2.3.2
pandas==2.0.3
ezdxf==1.1.0
shapely==2.0.1
Pillow==10.0.0